Rename util to http_util.
