+++ docs/source/api/http_util.rst
@@ -2,12 +2,12 @@
 Utilities
 *********
 
-The `siphon.util` module contains utility code to support making requests
+The `siphon.http_util` module contains utility code to support making requests
 using HTTP.
 
 :mod:`siphon.util`
 ======================
 
-.. automodule:: siphon.util
+.. automodule:: siphon.http_util
    :members:
    :undoc-members:
