+++ docs/source/api/index.rst
@@ -13,4 +13,4 @@
    catalog
    metadata
    ncssdataset
-   util
+   http_util
