+++ siphon/__init__.py
@@ -3,4 +3,4 @@ from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
-__all__ = ['catalog', 'testing', 'util']
+__all__ = ['catalog', 'testing', 'http_util']
