+++ siphon/http_util.py
